% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.perc.table.R
\name{create.perc.table}
\alias{create.perc.table}
\title{A percentile function}
\usage{
create.perc.table(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function creates a percentile table given a data frame.
}
\examples{
create.perc.table()
}
\keyword{percentile}
\keyword{quantile}
\keyword{table}
